Macro dir_assert::assert_paths

macro_rules! assert_paths {
    ($actual: expr, $expected: expr) => { ... };
}

Recursively scan contents of two directories and find differences.

eg.:

#[test]
fn should_directories_be_equal() {
    assert_paths!("actual", "expected");
}

This macro will panic if directories "actual" and "expected" differ at any depth.

It can accept both file names and directory names as arguments.